Add MemoryStorageService for managing data compartments

+++ src/Api/ChatHistory_v1.cs
@@ -14,25 +14,26 @@ using Aire.Memory.Models;
 using Aire.Sdk.AspNetCore;
 using Aire.Sdk.Models.Chat;
 using Aire.Sdk.Auth;
-using Aire.Sdk.Azure;
 using System.Web.Http;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using Aire.Memory.Services;
+using Aire.Sdk.Auth.Extensions;
 
 namespace Aire.Memory.Api;
 
 public class ChatHistory_v1
 {
     private readonly BlobContainerClient _chatlogs;
-    private readonly ITableStorageService _tables;
+    private readonly MemoryStorageService _storageService;
     private readonly IJwtTokenService _jwt;
 
-    public ChatHistory_v1(BlobServiceClient blobs, ITableStorageService tables, IJwtTokenService jwt)
+    public ChatHistory_v1(BlobServiceClient blobs, MemoryStorageService storageService, IJwtTokenService jwt)
     {
         _chatlogs = blobs.GetBlobContainerClient(AireConstants.Blobs.ChatLogs);
         _chatlogs.CreateIfNotExists(publicAccessType: PublicAccessType.None);
 
-        _tables = tables;
+        _storageService = storageService;
         _jwt = jwt;
     }
 
@@ -45,6 +46,7 @@ public class ChatHistory_v1
     [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(List<ChatLogMetadata>), Description = "List of chat metadata objects")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Missing platform authentication")]
     public async Task<IActionResult> GetChatHistory(
         [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "v1/chat-history")] HttpRequest req,
         FunctionContext context)
@@ -56,7 +58,11 @@ public class ChatHistory_v1
         if (!_jwt.CheckAuthorization(auth, requiredScopes: AireScopes.ReadChatHistory))
             return new ForbiddenResult();
 
-        var query = await _tables.QueryAsync<ChatLogEntity>(x => x.PartitionKey == auth.UserId);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+        var query = await tables.QueryAsync<ChatLogEntity>(x => x.PartitionKey == auth.UserId);
 
         var logs = await query.ToListAsync();
         var list = logs.Select(x => new ChatLogMetadata
@@ -78,7 +84,7 @@ public class ChatHistory_v1
     [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(ChatLog), Description = "List of chat messages and chat state")]
     [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found.")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
-    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param or missing platform authentication")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
     public async Task<IActionResult> GetChatHistoryWithId(
         [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "v1/chat-history/{id}")] HttpRequest req,
@@ -95,7 +101,12 @@ public class ChatHistory_v1
         if (string.IsNullOrWhiteSpace(id))
             return new BadRequestResult();
 
-        var entity = await _tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+
+        var entity = await tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
         if (entity == null)
             return new NotFoundResult();
 
@@ -120,7 +131,7 @@ public class ChatHistory_v1
         Description = "User token")]
     [OpenApiRequestBody("application/json", typeof(ChatLog), Description = "List of chat messages and chat state", Required = true)]
     [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(ChatLogMetadata), Description = "Chat log metadata")]
-    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid body")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid body or missing platform authentication")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
     public async Task<IActionResult> PostChatHistory(
@@ -141,7 +152,12 @@ public class ChatHistory_v1
         var entity = new ChatLogEntity(auth.UserId);
         await entity.SaveToBlob(_chatlogs, chat, auth.UserKey);
 
-        var add = await _tables.UpsertAsync(entity);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+
+        var add = await tables.UpsertAsync(entity);
         if (!add)
             return new InternalServerErrorResult();
 
@@ -163,7 +179,7 @@ public class ChatHistory_v1
     [OpenApiRequestBody("application/json", typeof(ChatLog), Description = "List of chat messages and chat state", Required = true)]
     [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(ChatLogMetadata), Description = "Chat log metadata")]
     [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found")]
-    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid body or param")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid body or param, or missing platform authentication")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
     public async Task<IActionResult> PutChatHistory(
@@ -185,13 +201,18 @@ public class ChatHistory_v1
         if (chat == null)
             return new BadRequestResult();
 
-        var chatlog = await _tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+
+        var chatlog = await tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
         if (chatlog == null)
             return new NotFoundResult();
 
         await chatlog.SaveToBlob(_chatlogs, chat, auth.UserKey);
 
-        var update = await _tables.UpsertAsync(chatlog);
+        var update = await tables.UpsertAsync(chatlog);
         if (!update)
             return new InternalServerErrorResult();
 
@@ -213,6 +234,7 @@ public class ChatHistory_v1
     [OpenApiResponseWithoutBody(HttpStatusCode.NoContent, Description = "The chatlogs were removed successfully")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Missing platform authentication")]
     public async Task<IActionResult> DeleteChatHistory(
         [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "v1/chat-history")] HttpRequest req,
         FunctionContext context)
@@ -224,14 +246,19 @@ public class ChatHistory_v1
         if (!_jwt.CheckAuthorization(auth, requiredScopes: AireScopes.DeleteChatHistory))
             return new ForbiddenResult();
 
-        var query = await _tables.QueryAsync<ChatLogEntity>(x => x.PartitionKey == auth.UserId);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+
+        var query = await tables.QueryAsync<ChatLogEntity>(x => x.PartitionKey == auth.UserId);
         var history = await query.ToListAsync();
 
         foreach (var chat in history)
         {
             await _chatlogs.DeleteBlobIfExistsAsync(chat.Id());
 
-            var delete = await _tables.DeleteAsync(chat);
+            var delete = await tables.DeleteAsync(chat);
             if (!delete)
                 return new InternalServerErrorResult();
         }
@@ -248,7 +275,7 @@ public class ChatHistory_v1
     [OpenApiParameter("id", Description = "Chat log identifier", In = ParameterLocation.Path, Required = true)]
     [OpenApiResponseWithoutBody(HttpStatusCode.NoContent, Description = "The chatlog(s) removed successfully")]
     [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found")]
-    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param")]
+    [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param or missing platform authentication")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
     [OpenApiResponseWithoutBody(HttpStatusCode.Forbidden, Description = "Access denied")]
     public async Task<IActionResult> DeleteChatHistoryWithId(
@@ -266,13 +293,18 @@ public class ChatHistory_v1
         if (string.IsNullOrWhiteSpace(id))
             return new BadRequestResult();
 
-        var chat = await _tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
+        if (auth.Platform == null)
+            return new BadRequestResult();
+
+        var tables = await _storageService.GetTableStorageService(auth.Platform, req.GetTargetService());
+
+        var chat = await tables.RetrieveAsync<ChatLogEntity>(auth.UserId, id);
         if (chat == null)
             return new NotFoundResult();
 
         await _chatlogs.DeleteBlobIfExistsAsync(chat.Id());
 
-        var delete = await _tables.DeleteAsync(chat);
+        var delete = await tables.DeleteAsync(chat);
         if (!delete)
             return new InternalServerErrorResult();
